requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.0
torch-spline-conv>=0.2.1

# NLP and embeddings
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# LLM support
llama-index>=0.9.0
accelerate>=0.20.0

# Data processing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=0.8.0
rich>=13.0.0
tensorboard>=2.13.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.22.0
